+++ src/crawler.cc
@@ -44,7 +44,7 @@ int main ( )
     for ( ; it != g_seeds.end(); it++) {
         thread::Task * task = 
             new thread::Task(handleUrl, (void *)strdup((*it).c_str()));
-        g_threadpool.addTask(task);
+        g_urltask_pool.addTask(task);
     }
 
     while (g_current_dnldcount < g_target_dnldcount) {
@@ -52,7 +52,6 @@ int main ( )
         do {
             nfds = epoll_wait(epollfd, events, MAX_EPOLL_EVENTS, -1);
             LOG(INFO) << "epoll_wait() returns nfds " << nfds << ", errno " << errno;
-            //fprintf(stderr, "tried epoll_wait(). nfds %d, errno %d\n", nfds, errno); 
         } while ((nfds < 0) && (EINTR == errno));
 
         for (int i = 0; i < nfds; i++) {
@@ -61,8 +60,6 @@ int main ( )
 
             LOG(INFO) << "events[" << i << "] event " << 
                     events[i].events << ", fd " << fdinfo->fd;
-            //fprintf(stderr, "events[%d] event %d, fd %d\n",
-            //        i, events[i].events, fdinfo->fd);
 
             int op = EPOLL_CTL_DEL;
             int sockfd = fdinfo->fd;
@@ -72,7 +69,7 @@ int main ( )
                 if ((ret = transport::recv(fdinfo)) != -1) {
                     thread::Task * task = 
                         new thread::Task(handleResponse, fdinfo);
-                    g_threadpool.addTask(task);
+                    g_responsetask_pool.addTask(task);
                     close(sockfd);
                 }
                 // handle uncomplete read (move out if want handle write as well)
@@ -81,7 +78,6 @@ int main ( )
                     op = EPOLL_CTL_MOD;
                     if (-1 == epoll_ctl(epollfd, op, fdinfo->fd, &events[i])) {
                         LOG(ERROR) << "epoll_ctl() failed. errno " << errno;
-                        //fprintf(stderr, "epoll_ctl() failed. errno %d\n", errno);
                         return -1;
                     }
                 }
@@ -90,7 +86,6 @@ int main ( )
                      (events[i].events & EPOLLHUP)) {
 
                 LOG(INFO) << "closing fd " << fdinfo->fd;
-                //fprintf(stderr, "closing fd %d\n", fdinfo->fd);
                 transport::freeFdInfo(fdinfo);
                 close(sockfd); // close() will remove fd from epoll set
             }
@@ -108,7 +103,6 @@ int init( )
     std::ifstream configfile(CONFIG_FILE);
     if (!configfile.is_open()) {
         std::cerr <<  "Cant open config file " << CONFIG_FILE;
-        //fprintf(stderr, "Can't open config file: %s\n", CONFIG.c_str());
         return -1;
     }
 
@@ -127,8 +121,6 @@ int init( )
         else {
             conf_map[prop] = value;
         }
-        LOG(DEBUG3) << prop << "=" << value; // DEBUG. to be removed
-        //fprintf(stderr, "%s=%s\n", prop.c_str(), value.c_str());
     }
 
     if (!conf_map[CONF_DOWNLOAD_DIR].empty()) {
@@ -153,6 +145,8 @@ int init( )
 
 int destroy( )
 {
+    g_urltask_pool.destroy();
+    g_responsetask_pool.destroy();
     crawl::destroy();
     logs::Logger::destroy();
 }
@@ -161,12 +155,10 @@ int destroy( )
 void handleUrl(void * arg )
 {
     LOG(INFO) << "entering task..";
-    //fprintf(stdout, "task executing handleUrl..\n");
     http::Url url((char *)arg);
     free(arg);
     if (url.getState() != http::OK) {
-        LOG(ERROR) << "parsing url failed: " << url.getState();
-        //fprintf(stdout, "url parsing failed: %d\n", url.getState());
+        LOG(WARNING) << "parsing url failed: " << url.getState();
         return;
     }
 
@@ -182,8 +174,7 @@ void handleUrl(void * arg )
     int ret = transport::send(fdinfo);
 
     if  (-1 == ret) {
-        LOG(ERROR) << "send request failed. errno " << errno;
-        //fprintf(stdout, "handleUrl. send failed. errno %d\n", errno);
+        LOG(WARNING) << "send request failed. errno " << errno;
         transport::freeFdInfo(fdinfo);
         close(sockfd);
         return;
@@ -202,7 +193,6 @@ void handleUrl(void * arg )
 void handleResponse(void * arg )
 {
     LOG(INFO) << "entering task..";
-    //fprintf(stdout, "task executing handleResponse..\n");
     struct transport::FdInfo * fdinfo = (struct transport::FdInfo *)arg;
     http::Url url(fdinfo->uri, true);
     http::Response resp(fdinfo->buff);
@@ -217,27 +207,17 @@ void handleResponse(void * arg )
     std::list<std::string> link_tbl;
     int total_links = crawl::reapLink(resp, link_tbl);
     LOG(INFO) << "crawler reaped " << total_links << " links";
-    //fprintf(stdout, "crawler reaped %d links\n", total_links);
     if (total_links <= 0) {
         return;
     }
     
     int valid_links = crawl::processLink(url, link_tbl);
     LOG(INFO) << "crawler filtered out " << total_links - valid_links << " links";
-    //fprintf(stdout, "crawler filtered out %d links\n", 
-    //    total_links - valid_links);
 
     std::list<std::string>::iterator it = link_tbl.begin();
     for ( ; it != link_tbl.end(); it++) {
         thread::Task * task = new thread::Task(handleUrl, (void *)strdup((*it).c_str()));
-        g_threadpool.addTask(task);
+        g_urltask_pool.addTask(task);
     }
 }
 
-            //  if (events[i].events & EPOLLOUT) {
-            //    if ((ret = transport::send(fdinfo)) != -1) {
-            //        events[i].events = EPOLLIN | EPOLLONESHOT;
-            //        op = EPOLL_CTL_MOD;
-            //    }
-            //}
-            //else